# voice-customer-service-lab/packages/contracts/openapi/voice-api.v1.json
{
  "components": {
    "schemas": {
...
      "def-18": {
        "additionalProperties": false,
        "properties": {
          "accepted": {
            "enum": [
              true
            ],
            "type": "boolean"
          },
          "replayed_count": {
            "maximum": 4,
            "minimum": 0,
            "type": "integer"
          },
          "tool_call_count": {
            "maximum": 4,
            "minimum": 1,
            "type": "integer"
          }
        },
        "required": [
          "accepted",
          "tool_call_count",
          "replayed_count"
        ],
        "title": "FunctionCallCallbackAck",
        "type": "object"
      },
      "def-19": {
        "additionalProperties": false,
        "properties": {
          "arguments": {
            "additionalProperties": false,
            "properties": {
              "order_reference": {
                "maxLength": 32,
                "minLength": 6,
                "pattern": "^[A-Za-z0-9-]+$",
                "type": "string"
              }
            },
            "required": [
              "order_reference"
            ],
            "type": "object"
          },
          "name": {
            "enum": [
              "get_order_status"
            ],
            "type": "string"
          },
          "tool_call_id": {
            "maxLength": 128,
            "minLength": 6,
            "pattern": "^call_[A-Za-z0-9_-]+$",
            "type": "string"
          }
        },
        "required": [
          "tool_call_id",
          "name",
          "arguments"
        ],
        "title": "MockBusinessToolCallRequest",
        "type": "object"
      },
      "def-2": {
        "additionalProperties": false,
        "properties": {
          "text": {
            "maxLength": 2000,
            "minLength": 1,
            "pattern": "\\S",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "AiDebugTurnRequest",
        "type": "object"
      },
      "def-20": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "enum": [
              "get_order_status"
            ],
            "type": "string"
          },
          "replayed": {
            "type": "boolean"
          },
          "result": {
            "$ref": "#/components/schemas/def-26"
          },
          "tool_call_id": {
            "maxLength": 128,
            "minLength": 6,
            "pattern": "^call_[A-Za-z0-9_-]+$",
            "type": "string"
          }
        },
        "required": [
          "tool_call_id",
          "name",
          "replayed",
          "result"
        ],
        "title": "MockBusinessToolCallResponse",
        "type": "object"
      },
...
      "def-3": {
        "additionalProperties": false,
        "properties": {
          "answer_mode": {
            "anyOf": [
              {
                "enum": [
                  "grounded_answer"
                ],
                "type": "string"
              },
              {
                "enum": [
                  "clarify"
                ],
                "type": "string"
              },
              {
                "enum": [
                  "direct_answer"
                ],
                "type": "string"
              },
              {
                "enum": [
                  "abstain"
                ],
                "type": "string"
              },
              {
                "enum": [
                  "tool_required"
                ],
                "type": "string"
              },
              {
                "enum": [
                  "handoff"
                ],
                "type": "string"
              },
              {
                "enum": [
                  "out_of_scope"
                ],
                "type": "string"
              },
              {
                "enum": [
                  "safety_refusal"
                ],
                "type": "string"
              }
            ]
          },
          "citations": {
            "items": {
              "additionalProperties": false,
              "properties": {
                "source_id": {
                  "maxLength": 128,
                  "minLength": 3,
                  "type": "string"
                },
                "title": {
                  "maxLength": 200,
                  "minLength": 1,
                  "type": "string"
                },
                "version": {
                  "maxLength": 64,
                  "minLength": 1,
                  "type": "string"
                }
              },
              "required": [
                "source_id",
                "title",
                "version"
              ],
              "type": "object"
            },
            "maxItems": 12,
            "type": "array"
          },
          "command_replayed": {
            "type": "boolean"
          },
          "evidence_status": {
            "anyOf": [
              {
                "enum": [
                  "sufficient"
                ],
                "type": "string"
              },
              {
                "enum": [
                  "none"
                ],
                "type": "string"
              },
              {
                "enum": [
                  "conflicting"
                ],
                "type": "string"
              },
              {
                "enum": [
                  "stale"
                ],
                "type": "string"
              },
              {
                "enum": [
                  "not_applicable"
                ],
                "type": "string"
              }
            ]
          },
          "execution": {
            "additionalProperties": false,
            "properties": {
              "model": {
                "anyOf": [
                  {
                    "maxLength": 64,
                    "minLength": 1,
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "model_usage": {
                "anyOf": [
                  {
                    "additionalProperties": false,
                    "properties": {
                      "input_tokens": {
                        "minimum": 0,
                        "type": "integer"
                      },
                      "output_tokens": {
                        "minimum": 0,
                        "type": "integer"
                      }
                    },
                    "required": [
                      "input_tokens",
                      "output_tokens"
                    ],
                    "type": "object"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "policy_version": {
                "maxLength": 128,
                "minLength": 3,
                "type": "string"
              },
              "provider_request_id": {
                "anyOf": [
                  {
                    "maxLength": 256,
                    "minLength": 1,
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "retriever": {
                "anyOf": [
                  {
                    "maxLength": 64,
                    "minLength": 1,
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "route_reason": {
                "maxLength": 64,
                "minLength": 2,
                "type": "string"
              },
              "router": {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              }
            },
            "required": [
              "policy_version",
              "router",
              "retriever",
              "model",
              "route_reason",
              "provider_request_id",
              "model_usage"
            ],
            "type": "object"
          },
          "round_id": {
            "maxLength": 80,
            "minLength": 8,
            "pattern": "^[a-z]{3,6}_[A-Za-z0-9]+$",
            "type": "string"
          },
          "schema_version": {
            "enum": [
              1
            ],
            "type": "number"
          },
          "session_id": {
            "maxLength": 80,
            "minLength": 8,
            "pattern": "^[a-z]{3,6}_[A-Za-z0-9]+$",
            "type": "string"
          },
          "spoken_text": {
            "maxLength": 4000,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "command_replayed",
          "session_id",
          "round_id",
          "answer_mode",
          "evidence_status",
          "spoken_text",
          "citations",
          "execution"
        ],
        "title": "AiDebugTurnResponse",
        "type": "object"
      },
...
      "def-37": {
        "additionalProperties": true,
        "properties": {
          "AppId": {
            "maxLength": 64,
            "minLength": 1,
            "type": "string"
          },
          "Message": {
            "maxLength": 32768,
            "minLength": 2,
            "type": "string"
          },
          "RoomID": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "Signature": {
            "maxLength": 512,
            "minLength": 1,
            "type": "string"
          },
          "TaskID": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "TaskType": {
            "enum": [
              "voiceChat"
            ],
            "type": "string"
          },
          "Type": {
            "enum": [
              "tool_calls"
            ],
            "type": "string"
          }
        },
        "required": [
          "Message",
          "Signature",
          "Type",
          "RoomID",
          "TaskID",
          "TaskType",
          "AppId"
        ],
        "title": "VolcengineFunctionCallCallbackRequest",
        "type": "object"
      },
...
      "def-39": {
        "contentMediaType": "text/event-stream",
        "title": "AiDebugStreamResponse",
        "type": "string"
      },
...
      "def-7": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "additionalProperties": false,
            "properties": {
              "code": {
                "pattern": "^[A-Z][A-Z0-9_]*$",
                "type": "string"
              },
              "correlation_id": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "retryable": {
                "type": "boolean"
              }
            },
            "required": [
              "code",
              "message",
              "retryable",
              "correlation_id"
            ],
            "title": "ApiErrorDetail",
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Voice Customer Service API",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/sessions/{session_id}/ai/debug-streams": {
      "post": {
        "operationId": "createAiDebugStream",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "maxLength": 80,
              "minLength": 8,
              "pattern": "^[a-z]{3}_[A-Za-z0-9]+$",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/def-2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/def-39"
                }
              }
            },
            "description": "Server-Sent Events stream"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "tags": [
          "ai-debug"
        ]
      }
    },
    "/api/v1/sessions/{session_id}/ai/debug-turns": {
      "post": {
        "operationId": "createAiDebugTurn",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "maxLength": 80,
              "minLength": 8,
              "pattern": "^[a-z]{3}_[A-Za-z0-9]+$",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/def-2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            },
            "description": "Default Response"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-3"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "tags": [
          "ai-debug"
        ]
      }
    },
...
    "/api/v1/sessions/{session_id}/mock-tool-calls": {
      "post": {
        "operationId": "invokeMockBusinessTool",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "maxLength": 80,
              "minLength": 8,
              "pattern": "^[a-z]{3}_[A-Za-z0-9]+$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/def-19"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-20"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "tags": [
          "mock-tools"
        ]
      }
    },
...
    "/internal/provider-callbacks/volcengine/function-calls": {
      "post": {
        "operationId": "receiveVolcengineFunctionCalls",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/def-37"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-18"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-7"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "tags": [
          "provider-callbacks"
        ]
      }
    }
  }
}